+++ ios/tinqle/AppDelegate.mm
@@ -59,13 +59,13 @@ didReceiveNotificationResponse:(UNNotificationResponse *)response
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
   
-  return [super application:application didFinishLaunchingWithOptions:launchOptions];
+//  return [super application:application didFinishLaunchingWithOptions:launchOptions];
   
   //for notofication
   UNUserNotificationCenter *center = [UNUserNotificationCenter currentNotificationCenter];
   center.delegate = self;
   
-  return YES;
+  return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }
 
 // kakao
